chore: improve InputElement exports
